DDF-3659 Work off review comments (3)
