+++ catalog/ui/catalog-ui-search/src/main/java/org/codice/ddf/catalog/ui/forms/data/ShareableMetacard.java
@@ -13,6 +13,9 @@
  */
 package org.codice.ddf.catalog.ui.forms.data;
 
+import static org.codice.ddf.catalog.ui.forms.data.ShareableAttributes.SHAREABLE_METADATA;
+import static org.codice.ddf.catalog.ui.forms.data.ShareableAttributes.SHAREABLE_TAG;
+
 import com.google.common.collect.Sets;
 import ddf.catalog.data.Attribute;
 import ddf.catalog.data.Metacard;
@@ -26,17 +29,25 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-/** A metacard that supports sharing. */
-public class ShareableMetacardImpl extends MetacardImpl {
-  public ShareableMetacardImpl(MetacardType type) {
+/**
+ * A metacard that supports sharing.
+ *
+ * <p><i>This code is experimental. While it is functional and tested, it may change or be removed
+ * in a future version of the library.</i>
+ *
+ * <p>TODO DDF-3671 Revisit sharing functionality for metacards
+ */
+@SuppressWarnings("squid:S1135" /* Action to-do has a ticket number and will be addressed later */)
+public class ShareableMetacard extends MetacardImpl {
+  public ShareableMetacard(MetacardType type) {
     super(type);
   }
 
-  public ShareableMetacardImpl(Metacard metacard) {
+  public ShareableMetacard(Metacard metacard) {
     super(metacard);
   }
 
-  public ShareableMetacardImpl(Metacard metacard, MetacardType type) {
+  public ShareableMetacard(Metacard metacard, MetacardType type) {
     super(metacard, type);
   }
 
@@ -47,13 +58,12 @@ public class ShareableMetacardImpl extends MetacardImpl {
    * @return true if the provided metacard is a shareable metacard, false otherwise.
    */
   public static boolean isShareableMetacard(Metacard metacard) {
-    return metacard != null
-        && metacard.getTags().stream().anyMatch(FormAttributes.Sharing.NAME::equals);
+    return metacard != null && metacard.getTags().stream().anyMatch(SHAREABLE_TAG::equals);
   }
 
   /** Wrap any metacard as a ShareableMetacardImpl. */
-  public static ShareableMetacardImpl from(Metacard metacard) {
-    return new ShareableMetacardImpl(metacard);
+  public static ShareableMetacard from(Metacard metacard) {
+    return new ShareableMetacard(metacard);
   }
 
   /**
@@ -89,10 +99,10 @@ public class ShareableMetacardImpl extends MetacardImpl {
   }
 
   public Set<String> getSharing() {
-    return new HashSet<>(getValues(FormAttributes.Sharing.FORM_SHARING));
+    return new HashSet<>(getValues(SHAREABLE_METADATA));
   }
 
   public void setSharing(Set<String> sharing) {
-    setAttribute(FormAttributes.Sharing.FORM_SHARING, new ArrayList<>(sharing));
+    setAttribute(SHAREABLE_METADATA, new ArrayList<>(sharing));
   }
 }
