+++ catalog/ui/catalog-ui-search/src/main/java/org/codice/ddf/catalog/ui/util/AccessUtil.java
@@ -23,10 +23,10 @@ import javax.annotation.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class AccessUtils {
-  private static final Logger LOGGER = LoggerFactory.getLogger(AccessUtils.class);
+public class AccessUtil {
+  private static final Logger LOGGER = LoggerFactory.getLogger(AccessUtil.class);
 
-  private AccessUtils() {
+  private AccessUtil() {
     // Should not be instantiated
   }
